+++ pkg/api/v1alpha1/cloudstackdatacenterconfig_types.go
@@ -35,8 +35,10 @@ type CloudStackDatacenterConfigSpec struct {
 }
 
 type CloudStackResourceIdentifier struct {
-	Id   string `json:"id"`
-	Name string `json:"name"`
+	// +optional
+	Id   string `json:"id,omitempty"`
+	// +optional
+	Name string `json:"name,omitempty"`
 }
 
 const (